test(amplify-e2e-tests): remove content from hosting bucket before del
Updated e2e test to remove the hosting bucket content before deleting the bucket
