+++ packages/amplify-e2e-core/package.json
@@ -26,6 +26,7 @@
     "graphql-transformer-core": "6.19.2",
     "jest-environment-node": "^25.1.0",
     "node-pty-prebuilt-multiarch": "^0.9.0",
+    "lodash": "^4.17.15",
     "retimer": "2.0.0",
     "rimraf": "^3.0.0",
     "strip-ansi": "^6.0.0",
