+++ packages/amplify-e2e-core/tsconfig.json
@@ -1,5 +1,6 @@
 {
   "compilerOptions": {
+    "composite": true,
     "skipLibCheck": true,
     "allowJs": true,
     "target": "es6",
