+++ packages/amplify-e2e-tests/src/__tests__/delete.test.ts
@@ -9,6 +9,8 @@ import {
   getAmplifyConfigIOSPath,
   getAWSConfigAndroidPath,
   getAmplifyConfigAndroidPath,
+  bucketNotExists,
+  deleteS3Bucket,
 } from 'amplify-e2e-core';
 import { addEnvironment, checkoutEnvironment, removeEnvironment } from '../environment/add-env';
 import { addApiWithoutSchema } from 'amplify-e2e-core';
@@ -90,7 +92,7 @@ describe('amplify delete', () => {
     const meta = amplifyMeta.providers.awscloudformation;
     const bucketName = meta.DeploymentBucketName;
     expect(await bucketExists(bucketName)).toBeTruthy();
-    await deleteBucket(bucketName);
+    await deleteS3Bucket(bucketName);
     await deleteProject(projRoot);
   });
 });
@@ -168,61 +170,4 @@ async function timeout(timeout: number) {
   return new Promise((resolve, reject) => {
     setTimeout(resolve, timeout);
   });
-}
-
-async function bucketNotExists(bucket: string) {
-  const s3 = new S3();
-  const params = {
-    Bucket: bucket,
-    $waiter: { maxAttempts: 10, delay: 30 },
-  };
-  try {
-    await s3.waitFor('bucketNotExists', params).promise();
-    return true;
-  } catch (error) {
-    if (error.statusCode === 200) {
-      return false;
-    }
-    throw error;
-  }
-}
-
-async function deleteBucket(bucket: string) {
-  const s3 = new S3();
-  let continuationToken = null;
-  const objectKey = [];
-  let truncated = false;
-  do {
-    const results = await s3
-      .listObjectsV2({
-        Bucket: bucket,
-        ContinuationToken: continuationToken,
-      })
-      .promise();
-    results.Contents.forEach(r => {
-      objectKey.push({ Key: r.Key });
-    });
-
-    continuationToken = results.NextContinuationToken;
-    truncated = results.IsTruncated;
-  } while (truncated);
-  const chunkedResult = _.chunk(objectKey, 1000);
-  const deleteReq = chunkedResult
-    .map(r => {
-      return {
-        Bucket: bucket,
-        Delete: {
-          Objects: r,
-          Quiet: true,
-        },
-      };
-    })
-    .map(delParams => s3.deleteObjects(delParams).promise());
-  await Promise.all(deleteReq);
-  await s3
-    .deleteBucket({
-      Bucket: bucket,
-    })
-    .promise();
-  await bucketNotExists(bucket);
 }
