+++ packages/amplify-e2e-tests/tsconfig.json
@@ -12,5 +12,8 @@
     "declaration": true,
     "typeRoots": ["../../node_modules/@types", "node_modules/@types", "./typings"]
   },
+  "references": [
+    {"path": "../amplify-e2e-core"},
+  ],
   "exclude": ["node_modules", "lib", "__tests__"]
 }
